pyproject.toml
[project]
name = "tonecrafterai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.14.0",
    "ddgs>=9.10.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.0",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.10",
    "llama-parse>=0.6.94",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]
